아이트래킹 추가
그냥 러프하게 합친거라 최적화 같은건 안되어 있음. 그냥 써도 되면 시간 관계상 건들지 말고 리소스 모자라 보이면 제대로 하는 방향으로 진행.
