+++ requirements.txt
@@ -1,0 +1,8 @@
+numpy == 1.16.1
+opencv_python == 3.4.5.20
+dlib == 19.21.1
+opencv-python == 4.5.1.48
+opencv-contrib-python == 4.5.1.48
+Tkinter (Tkinter for python 2.7, tkinter for python 3.x)
+face_recognition == 1.3.0
+face_recognition-models == 0.3.0
